+++ src/main/java/exception/FrankUnknownException.java
@@ -1,4 +1,4 @@
-package Exception;
+package exception;
 
 public class FrankUnknownException extends FrankException {
     public FrankUnknownException() {
