+++ src/main/java/task/Event.java
@@ -1,4 +1,4 @@
-package Task;
+package task;
 
 public class Event extends Task {
