+++ src/main/java/task/Task.java
@@ -1,4 +1,4 @@
-package Task;
+package task;
 
 public class Task {
     protected String description;
