+++ src/main/java/task/Todo.java
@@ -1,4 +1,4 @@
-package Task;
+package task;
 
 public class Todo extends Task {
     public Todo(String description) {
